requirements.txt
openai==1.30.1
streamlit==1.35.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
plotly==5.22.0
tenacity==8.3.0
